Ecommerce_Website/requirements.txt
colorama           
cryptography        
Django               
django-cleanup       
django-phonenumber-field 
django-widget-tweaks     
oauthlib              
oscrypto             
phonenumbers    
pillow         
pip               
python-dotenv      
PyYAML            
reportlab            
requests               
requests-oauthlib       
social-auth-app-django  
social-auth-core        
sqlparse                
sslcommerz_python_api   
whitenoise